cmake: add simple CTest/TAP testing infrastructure
Tests are located in the tests/ directory and registered with CMake.
Tests can then be run by ctest, e.g.:
	ctest --output-on-failure

This works, but is not optimal, because ctest considers each test
executable as a unit. So either the executable exits successfully
or it doesn't. This is suboptimal, because often having multiple
related tests in the same file is desired.
Address this by providing a tap.h header that implements a simple
Perl Test::More like API and yield TAP output. TAP can be colorized
by the often installed prove(1) utility.

As the Test executable directory is named appropriately, running
prove -v in the CMake build directory should suffice.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>